Created AuthContext and added auth context to all routes including navbar
