+++ client/src/App.jsx
@@ -8,25 +8,24 @@ import ForgetPassword from "./pages/ForgetPassword";
 import DashboardPage from "./pages/DashboardPage";
 import Navbar from "./components/Navbar";
 import ViewRecipeDetails from "./pages/viewRecipeDetails";
-
+import AuthProvider from "./context/AuthContext";
 
 function App() {
   return (
     <Router>
-    <Navbar />
-    
-      <Routes>
-        <Route path="/" exact element={<Login />} />
-        <Route path="/registration" exact element={<SignUp />} />
-        <Route path="/forgetPassword" exact element={<ForgetPassword />} />
-        <Route path="/dashboard" exact element={<DashboardPage />} />
-        <Route path="/explore" exact element={<Explore />} />
-        <Route path="/result/:term" exact element={<SearchResult />} /> 
-        <Route path="/result/:term" element={<SearchResult />} />
-        <Route path="/viewRecipeDetails/:recipeID" element={<ViewRecipeDetails />} />
-      </Routes>
+      <AuthProvider>
+        <Navbar />
+        <Routes>
+          <Route path="/login" exact element={<Login />} />
+          <Route path="/registeration" exact element={<SignUp />} />
+          <Route path="/forgetPassword" exact element={<ForgetPassword />} />
+          <Route path="/dashboard" exact element={<DashboardPage />} />
+          <Route path="/explore" exact element={<Explore />} />
+          <Route path="/result/:term" exact element={<SearchResult />} />
+          <Route path="/viewRecipeDetails/:recipeID" element={<ViewRecipeDetails />} />
+        </Routes>
+      </AuthProvider>
     </Router>
-
   );
 }
